+++ processor/memorylimiterprocessor/memorylimiter_test.go
@@ -116,11 +116,7 @@ func TestMetricsMemoryPressureResponse(t *testing.T) {
 		readMemStatsFn: func(ms *runtime.MemStats) {
 			ms.Alloc = currentMemAlloc
 		},
-		obsrep: obsreport.NewProcessor(obsreport.ProcessorSettings{
-			Level:       configtelemetry.LevelNone,
-			ProcessorID: config.NewComponentID(typeStr),
-		}),
-
+		obsrep: newObsReport(),
 		logger: zap.NewNop(),
 	}
 	mp, err := processorhelper.NewMetricsProcessor(
@@ -189,10 +185,7 @@ func TestTraceMemoryPressureResponse(t *testing.T) {
 		readMemStatsFn: func(ms *runtime.MemStats) {
 			ms.Alloc = currentMemAlloc
 		},
-		obsrep: obsreport.NewProcessor(obsreport.ProcessorSettings{
-			Level:       configtelemetry.LevelNone,
-			ProcessorID: config.NewComponentID(typeStr),
-		}),
+		obsrep: newObsReport(),
 		logger: zap.NewNop(),
 	}
 	tp, err := processorhelper.NewTracesProcessor(
@@ -261,10 +254,7 @@ func TestLogMemoryPressureResponse(t *testing.T) {
 		readMemStatsFn: func(ms *runtime.MemStats) {
 			ms.Alloc = currentMemAlloc
 		},
-		obsrep: obsreport.NewProcessor(obsreport.ProcessorSettings{
-			Level:       configtelemetry.LevelNone,
-			ProcessorID: config.NewComponentID(typeStr),
-		}),
+		obsrep: newObsReport(),
 		logger: zap.NewNop(),
 	}
 	lp, err := processorhelper.NewLogsProcessor(
@@ -453,4 +443,14 @@ func TestBallastSizeMiB(t *testing.T) {
 			assert.Equal(t, tt.expectResult, tt.expectedMemLimiterBallastSize*mibBytes == ballastExt.(*ballastextension.MemoryBallast).GetBallastSize())
 		})
 	}
+}
+
+func newObsReport() *obsreport.Processor {
+	set := obsreport.ProcessorSettings{
+		ProcessorID:             config.NewComponentID(typeStr),
+		ProcessorCreateSettings: componenttest.NewNopProcessorCreateSettings(),
+	}
+	set.ProcessorCreateSettings.MetricsLevel = configtelemetry.LevelNone
+
+	return obsreport.NewProcessor(set)
 }